Machine learning/Assignment Week 11/requirements.txt
appdirs==1.4.4
asteval==0.9.18
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
biopython==1.77
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
deepdiff==5.0.1
et-xmlfile==1.0.1
fs==2.4.11
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
isaExplorer==0.1.0
isatools==0.11.0
iso8601==0.1.12
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
kiwisolver==1.2.0
lmfit==1.0.1
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mzml2isa==1.0.3
networkx==2.4
nPYc==1.2.3
numpy==1.19.0
openpyxl==2.6.4
ordered-set==4.0.2
pandas==1.0.5
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
plotly==4.8.2
progressbar2==3.51.4
prompt-toolkit==3.0.5
pronto==0.12.2
ptyprocess==0.6.0
pyChemometrics==0.1
Pygments==2.6.1
pyopenms==2.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
six==1.15.0
soupsieve==2.0.1
statsmodels==0.11.1
svgwrite==1.4
tabulate==0.8.7
threadpoolctl==2.1.0
traitlets==4.3.3
uncertainties==3.1.4
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.1.0
